Add edit-link to card when logged in, styled card and navbar
