--- src/components/Agencies/Agencies.js
@@ -1,51 +1,0 @@
-import React from "react";
-import { MdLocationCity, MdWebAsset, MdLocationOn } from "react-icons/md";
-import "./Agencies.css";
-
-export default function Agencies({ agency }) {
-  return (
-    <div className="main-agencies">
-      <div className="agency-card-container">
-        <div className="company-logo">
-          <img
-            src={agency.logo_image_src}
-            alt={`${agency.company_name} company logo`}
-          />
-        </div>
-        <div className="agency-card-info" key={agency.company_id}>
-          <div className="company-name">
-            <h2>{agency.company_name.replace("&amp;", "&")}</h2>
-          </div>
-          <div className="company-info">
-            <div>
-              <h5>
-                <span className="icon-gap">
-                  <MdLocationCity />
-                </span>
-                {agency.company_size} Employees
-              </h5>
-            </div>
-            <div>
-              <h5>
-                <span className="icon-gap">
-                  <MdLocationOn />
-                </span>
-                {agency.city_name}
-              </h5>
-            </div>
-            <div className="www-link">
-              <h5>
-                <span className="icon-gap">
-                  <MdWebAsset />
-                </span>
-                <a href={agency.website} target="_blank" rel="noreferrer">
-                  website
-                </a>
-              </h5>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
