+++ src/context/AuthContext.js
@@ -5,38 +5,38 @@ import { APIUrl } from "../config/config";
 export const AuthContext = createContext({});
 
 export default function AuthContextProvider({ children }) {
-  // const [name, setName] = useState("");
-  // const [email, setEmail] = useState("");
   const [username, setUsername] = useState("abczzz13");
   const [password, setPassword] = useState("v$*UfieAepz7zuYphJ^^");
   const [userToken, setUserToken] = useState("");
+  const [decoded, setDecoded] = useState({});
   const [redirect, setRedirect] = useState(false);
 
-  // username: "abczzz13",
-  // password: "v$*UfieAepz7zuYphJ^^",
-  const token =
-    "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c";
-
-  const decoded = jwt_decode(userToken ? userToken.token : token);
-  console.log("decoded", decoded);
-  console.log("username-auth", username);
   console.log("userToken", userToken);
+  console.log("decoded", decoded);
 
   const login = async (e) => {
     e.preventDefault();
-    const response = await fetch(`${APIUrl}/token`, {
-      method: "POST",
-      headers: {
-        "Content-type": "application/json",
-      },
-      body: JSON.stringify({
-        username,
-        password,
-      }),
-    });
-    const content = await response.json();
-    setUserToken(content);
-    setRedirect(true);
+    try {
+      const response = await fetch(`${APIUrl}/token`, {
+        method: "POST",
+        headers: {
+          "Content-type": "application/json",
+          Authorization: "Bearer" + userToken.token,
+        },
+        body: JSON.stringify({
+          username,
+          password,
+        }),
+      });
+
+      const content = await response?.json();
+      setUserToken(content);
+      setRedirect(true);
+      const decoded = jwt_decode(content.token);
+      setDecoded(decoded);
+    } catch (e) {
+      console.error(e);
+    }
   };
 
   const logout = () => {
@@ -50,10 +50,6 @@ export default function AuthContextProvider({ children }) {
     setPassword,
     userToken,
     setUserToken,
-    // name,
-    // setName,
-    // email,
-    // setEmail,
     login,
     redirect,
     logout,
